Add files via upload
Fixed a critical implicit narrowing conversion bug in the float_precision::float_precision(long long i, ...) constructor in the fprecision.h file on line 666. Instead of calling the abs(i) function, which selected the 32-bit version of int abs(int) under certain GCC compilation conditions, the llabs(i) function is now used, which is designed to work with 64-bit long long integers. This prevents an incorrect narrowing conversion of the 64-bit value LLONG_MAX (9,223,372,036,854,775,807) to the 32-bit value -1, which was then converted to 1, and ensures that the modulus of large numbers is calculated correctly, ensuring that the library is cross-platform compatible.